Added validation for a new expense details, minor UX changes in layouts

+++ app/src/main/java/com/example/expensemonitoring/Room/Dao/Tuples.kt
@@ -5,5 +5,5 @@ import androidx.room.ColumnInfo
 data class CategoriesTuple(
     @ColumnInfo(name = "month") var categoryMonth: String,
     @ColumnInfo(name = "year") var categoryYear: String,
-    @ColumnInfo(name = "amount_sum") var categorySum: Long,
+    @ColumnInfo(name = "amount_sum") var categorySum: Double,
 )
